+++ DJet ECU Tester/PC Application/DJetronicECUTester/Status.cs
@@ -16,10 +16,10 @@ namespace DJetronicECUTester
         public UInt32 DwellAngle;
         public bool FuelPumpOn;
         public bool ColdStartOn;
-        public UInt32 PulseWidth_I;     // us
-        public UInt32 PulseWidth_II;    // us
-        public UInt32 PulseWidth_III;   // us
-        public UInt32 PulseWidth_IV;    // us
+        public UInt16 PulseWidth_I;     // us
+        public UInt16 PulseWidth_II;    // us
+        public UInt16 PulseWidth_III;   // us
+        public UInt16 PulseWidth_IV;    // us
         public UInt32 Pressure;
     };
 }
